Созданны сервисы парсинга и валидации. Json-файлы теперь сохраняют данные не в одну строчку, а в виде структуры
