+++ AdvertisingPlatforms.Domain/Interfaces/IFileParser.cs
@@ -1,0 +1,9 @@
+﻿using AdvertisingPlatforms.Domain.Models;
+
+namespace AdvertisingPlatforms.Domain.Interfaces
+{
+    public interface IFileParser
+    {
+        public DataFromFile? GetParseData(string fileContent);
+    }
+}
