+++ AdvertisingPlatforms.Domain/Interfaces/IFileValidator.cs
@@ -1,0 +1,7 @@
+﻿namespace AdvertisingPlatforms.Domain.Interfaces
+{
+    public interface IFileValidator
+    {
+        bool IsValid(string? data);
+    }
+}
