--- AdvertisingPlatforms.Domain/Interfaces/IParser.cs
@@ -1,7 +1,0 @@
-﻿namespace AdvertisingPlatforms.Domain.Interfaces
-{
-    public interface IParser
-    {
-        
-    }
-}
