--- AdvertisingPlatforms.Domain/Interfaces/IValidator.cs
@@ -1,7 +1,0 @@
-﻿namespace AdvertisingPlatforms.Domain.Interfaces
-{
-    public interface IValidator
-    {
-        bool IsValid();
-    }
-}
