+++ AdvertisingPlatforms.sln
@@ -14,8 +14,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdvertisingPlatforms.Tests"
 		{8ECAD98E-96DA-D532-CFDF-200731811F44} = {8ECAD98E-96DA-D532-CFDF-200731811F44}
 	EndProjectSection
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Элементы решения", "Элементы решения", "{754FC069-D67B-A9D7-50A1-8D1CA196D8F1}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdvertisingPlatforms.Domain", "AdvertisingPlatforms.Domain\AdvertisingPlatforms.Domain.csproj", "{2285FDCF-4C97-7B45-BCA4-1AF65EBDDA1A}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AdvertisingPlatforms.DAL", "AdvertisingPlatforms.DAL\AdvertisingPlatforms.DAL.csproj", "{8ECAD98E-96DA-D532-CFDF-200731811F44}"
