--- ReadMe.txt
@@ -1,5 +1,0 @@
-1) Для удобства тестирования запросов, при запуске выводится html-разметка
-
-2) Сервер содержит данные по умолчанию
-
-3) Тесты сгруппированны по методам (5 групп)
